+++ flowsa/data_source_scripts/Blackhurst_IO.py
@@ -10,7 +10,7 @@ Scrapes data from Blackhurst paper 'Direct and Indirect Water Withdrawals for US
 import tabula
 import io
 from flowsa.common import *
-from flowsa.flowbyfunctions import  assign_fips_location_system
+from flowsa.flowbyfunctions import assign_fips_location_system
 
 
 # Read pdf into list of DataFrame
@@ -50,7 +50,9 @@ def bh_parse(dataframe_list, args):
 def convert_blackhurst_data_to_gal_per_year(df, attr):
 
     import flowsa
-    from flowsa.flowbyfunctions import clean_df, fba_fill_na_dict, harmonize_units
+    from flowsa.flowbyfunctions import fba_fill_na_dict
+    from flowsa.dataclean_functions import harmonize_units
+    from flowsa.dataclean_functions import clean_df
 
     # load the bea make table
     bmt = flowsa.getFlowByActivity(datasource='BEA_Make_Table',
@@ -84,7 +86,9 @@ def convert_blackhurst_data_to_gal_per_employee(df_wsec, attr, method):
 
     import flowsa
     from flowsa.mapping import add_sectors_to_flowbyactivity
-    from flowsa.flowbyfunctions import clean_df, fba_fill_na_dict, filter_by_geoscale, harmonize_units
+    from flowsa.flowbyfunctions import fba_fill_na_dict, filter_by_geoscale
+    from flowsa.dataclean_functions import harmonize_units
+    from flowsa.dataclean_functions import clean_df
     from flowsa.fbs_allocation_functions import proportional_allocation_by_location_and_activity
     from flowsa.data_source_scripts.BLS_QCEW import clean_bls_qcew_fba
 
@@ -146,7 +150,7 @@ def scale_blackhurst_results_to_usgs_values(df_to_scale, attr):
     """
 
     import flowsa
-    from flowsa.flowbyfunctions import harmonize_units
+    from flowsa.dataclean_functions import harmonize_units
 
     # determine national level published withdrawal data for usgs mining in FBS method year
     pv_load = flowsa.getFlowByActivity(datasource="USGS_NWIS_WU", year=str(attr['helper_source_year']),
