+++ flowsa/data_source_scripts/EIA_MECS.py
@@ -4,8 +4,8 @@
 
 import io
 from flowsa.common import *
-from flowsa.flowbyfunctions import assign_fips_location_system, replace_strings_with_NoneType, \
-    replace_NoneType_with_empty_cells
+from flowsa.flowbyfunctions import assign_fips_location_system
+from flowsa.dataclean_functions import replace_strings_with_NoneType, replace_NoneType_with_empty_cells
 import yaml
 
 """
