+++ scripts/FlowByActivity_Datasets/write_FBA_BEA_Make.py
@@ -9,7 +9,7 @@ Generation of BEA Make data as FBA
 from flowsa.common import *
 import pandas as pd
 from flowsa.flowbyactivity import store_flowbyactivity
-from flowsa.flowbyfunctions import add_missing_flow_by_fields
+from flowsa.dataclean_functions import add_missing_flow_by_fields
 
 year = 2012
 level = 'Detail'
