+++ tests/test_flowbysector_method4.py
@@ -21,7 +21,7 @@ class TestFlowBySectorFunctions(unittest.TestCase):
         flows = aggregator(flows, fbs_default_grouping_fields)
         totaltotaldomestic = flows['FlowAmount'].loc[(flows['Flowable'] == 'total') &
                                                      (flows['Context'] == 'total') &
-                                                     (flows['SectorProducedBy'] == 'None') &
+                                                     (flows['SectorProducedBy'] is None) &
                                                      (flows['SectorConsumedBy'] == 'F010')].reset_index(drop=True)
         totaltotalpstodomestic = flows['FlowAmount'].loc[(flows['Flowable'] == 'total') &
                                                      (flows['Context'] == 'total') &
